requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==41.0.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-async-sqlalchemy==0.6.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.20.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
priority==2.0.0
psutil==5.9.7
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.16.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyzmq==25.1.2
requests==2.31.0
rich==13.7.0
roundrobin==0.0.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
